feat: move `FeedImageDataLoaderWithFallbackComposite` to production
